+++ scripts/vscode-Install-Extension.ps1
@@ -1,2 +1,2 @@
 # Vscode install vsix
-Invoke-Expression "code --install-extension smartopen-1.3.2.vsix"
+Invoke-Expression "code --install-extension smartopen-1.4.1.vsix"